requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
SQLAlchemy
PyMySQL
python-dotenv
cryptography